requirements.txt
django
djangorestframework
environs
djangorestframework-simplejwt
django-allauth
dj-rest-auth[with_social]
django-cors-headers
drf-yasg
beautifulsoup4

